implemented ElementWriter and PageElementWriter
