+++ src/_elementWriter.js
@@ -1,80 +1,11 @@
 import {isNumber, pack, offsetVector} from './helpers';
-import Line from './line';
-import DocumentContext from './documentContext';
 
 /**
  * A line/vector writer, which adds elements to current page
  * and sets their positions based on the context
  */
 class ElementWriter {
-	constructor(context, tracker) {
-		this.context = context;
-		this.contextStack = [];
-		this.tracker = tracker;
-	}
 
-	addLine(line, dontUpdateContextPosition, index) {
-		let height = line.getHeight();
-		let context = this.context;
-		let page = context.getCurrentPage();
-		let position = this.getCurrentPositionOnPage();
-
-		if (context.availableHeight < height || !page) {
-			return false;
-		}
-
-		line.x = context.x + (line.x || 0);
-		line.y = context.y + (line.y || 0);
-
-		this.alignLine(line);
-
-		addPageItem(page, {
-			type: 'line',
-			item: line
-		}, index);
-		this.tracker.emit('lineAdded', line);
-
-		if (!dontUpdateContextPosition) {
-			context.moveDown(height);
-		}
-
-		return position;
-	}
-
-	alignLine(line) {
-		let width = this.context.availableWidth;
-		let lineWidth = line.getWidth();
-
-		let alignment = line.inlines && line.inlines.length > 0 && line.inlines[0].alignment;
-
-		let offset = 0;
-		switch (alignment) {
-			case 'right':
-				offset = width - lineWidth;
-				break;
-			case 'center':
-				offset = (width - lineWidth) / 2;
-				break;
-		}
-
-		if (offset) {
-			line.x = (line.x || 0) + offset;
-		}
-
-		if (alignment === 'justify' &&
-			!line.newLineForced &&
-			!line.lastLineInParagraph &&
-			line.inlines.length > 1) {
-			let additionalSpacing = (width - lineWidth) / (line.inlines.length - 1);
-
-			for (let i = 1, l = line.inlines.length; i < l; i++) {
-				offset = i * additionalSpacing;
-
-				line.inlines[i].x += offset;
-				line.inlines[i].justifyShift = additionalSpacing;
-			}
-		}
-	}
 
 	addImage(image, index) {
 		let context = this.context;
@@ -206,16 +137,11 @@ class ElementWriter {
 	}
 
 	addFragment(block, useBlockXOffset, useBlockYOffset, dontUpdateContextPosition) {
-		let ctx = this.context;
-		let page = ctx.getCurrentPage();
 
-		if (!useBlockXOffset && block.height > ctx.availableHeight) {
-			return false;
-		}
 
 		block.items.forEach((item) => {
 			switch (item.type) {
-				case 'line':
+			/*case 'line':
 					let l = cloneLine(item.item);
 
 					l.x = (l.x || 0) + (useBlockXOffset ? (block.xOffset || 0) : ctx.x);
@@ -226,6 +152,7 @@ class ElementWriter {
 						item: l
 					});
 					break;
+*/
 
 				case 'vector':
 					let v = pack(item.item);
@@ -251,61 +178,7 @@ class ElementWriter {
 			}
 		});
 
-		if (!dontUpdateContextPosition) {
-			ctx.moveDown(block.height);
-		}
-
-		return true;
 	}
-
-	/**
-	 * Pushes the provided context onto the stack or creates a new one
-	 *
-	 * pushContext(context) - pushes the provided context and makes it current
-	 * pushContext(width, height) - creates and pushes a new context with the specified width and height
-	 * pushContext() - creates a new context for unbreakable blocks (with current availableWidth and full-page-height)
-	 */
-	pushContext(contextOrWidth, height) {
-		if (contextOrWidth === undefined) {
-			height = this.context.getCurrentPage().height - this.context.pageMargins.top - this.context.pageMargins.bottom;
-			contextOrWidth = this.context.availableWidth;
-		}
-
-		if (isNumber(contextOrWidth)) {
-			contextOrWidth = new DocumentContext({width: contextOrWidth, height: height}, {left: 0, right: 0, top: 0, bottom: 0});
-		}
-
-		this.contextStack.push(this.context);
-		this.context = contextOrWidth;
-	}
-
-	popContext() {
-		this.context = this.contextStack.pop();
-	}
-
-	getCurrentPositionOnPage() {
-		return (this.contextStack[0] || this.context).getCurrentPosition();
-	}
-}
-
-function addPageItem(page, item, index) {
-	if (index === null || index === undefined || index < 0 || index > page.items.length) {
-		page.items.push(item);
-	} else {
-		page.items.splice(index, 0, item);
-	}
-}
-
-function cloneLine(line) {
-	let result = new Line(line.maxWidth);
-
-	for (let key in line) {
-		if (line.hasOwnProperty(key)) {
-			result[key] = line[key];
-		}
-	}
-
-	return result;
 }
 
 export default ElementWriter;
